Disallow merge when invoke-super resolves to private method
This CL prevents a class A from being merged into its subclass B if there is an invoke-super instruction that resolves to a private method in A. This is needed to ensure that the program will continue to yield an IllegalAccessError.

This case is caught by the test MemberResolutionTest.lookupPrivateSuperFromSubClass.

Change-Id: I3d8297e261fee6391f16c606d9368e18d0460e11