Access to most recent AppInfo and GraphLense via AppView
Note that this CL fixes a bug in PublicizerLense that manifests when vertical class merging is enabled. The lense only rewrote invoke-direct to invoke-virtual if it could find the targeted method on the holder of the method. However, when looking for the method, the lense did not consider the current state of the world. In particular, it failed to consider that the original holder of the method could have been merged into its subtype.

Change-Id: I0f25321720a5ebbe0aeb29690f4cd7f75f09c0d5