Add Proguard testing rules for controlling inlining
This adds two new rules -forceinline and -neverinling which is only
available for testing. This should help in getting the inlining
required for specific tests instead of having to use more indirect
methods to either ensure or prohibit inlining.

This CL does not change any existing tests to use these rules.

Bug: 110190426
Change-Id: Icdaf71e353b90f4c6ceb452ea3c49bc4948c7edd