Avoid vertical class merging if virtual method is hit by invoke-direct
It is valid to have an invoke-direct instruction in a default interface method that targets another default method in the same interface (see InterfaceMethodDesugaringTests.testInvokeSpecialToDefaultMethod). However, in a class, that would lead to a verification error. Therefore, this CL disallows merging such interfaces into their subtypes.

Note that, during vertical class merging, we create a private method for each virtual method that is moved to the subtype. Therefore, it would be possible to redirect the invoke-direct instructions that target a virtual method to invoke-direct instructions that target the newly created private methods. However, this would require maintaining potentially many mappings in the graph lense that would likely never be used in practice.

Change-Id: I24ec192c25d4357fa8582edffc6db80ff628e7e0