Keep assertion code when required
Android VMs (Dalvik and Art) does not support Java assertions that
the standard Java VM enables using the -ea option. So D8 and R8
strips the code for Java assertions is removed. However in some
cases this is not desired:

* For the class-file backend it should just be kept, as the
  output runs on a normal Java VM

* If the javac generated code for Java assertions has been
  rewritten through byte code rewriting to activate it
  through other means than passing -ea to the Java VM.

This change will always keep the javac generated assertions code
for the class-file backend and keep it for dex output if the code
does not follow the exact pattern (currently) generated by javac.

For d8 this changes compilation to always compile <clinit> for
a class before any other methods and use that as an indication
of wether the original javac generated assertions code is present.

Bug: 110887293
Change-Id: I27e1e7711b3acc6dea9df46adee9a32a7b24ee7e