InterfaceProcessor should keep some bridge methods
This CL updates InterfaceProcessor such that some default interface methods will be kept as virtual methods even if they are marked as bridge methods. This is needed if there is a default interface method that is marked as a bridge method, but does not override a method declared in a supertype of the current interface (such a situation can happen as a result of vertical class merging).

Change-Id: I60905948401f647a0ed27582e128ef562a4c8a84