Refine inlining constraints with target holder.
When a method has instructions with different targets, method's inlining
constraint is not simply a min of each instruction's Inliner.Constraint.
One regression example is http://b/111080693#comment6 (or go/r8g/23940).

This CL introduces ConstraintWithTarget, a wrapper of Constraint along
with its target holder, and tweak all uses of Constraint with it.
Calculation of min (or meet) of two ConstraintWithTarget take into
account target holders too so that constraints from unrelated items are
not incorrectly combined.

Bug: 111080693
Change-Id: Ic545b99a28ac1f97c3ff603ddb7824ad68111a0d