Strip invalid locals information instead of failing compilation.
The LocalVariableTable in a class file can be completely invalid.
This typically happens when there is bytecode instrumentation in
play. Instead of failing with an exception, we typically bail out
and then just strip the locals information from the method when
it doesn't make sense.

This change extends that approach to situations where the start
or end point of a local variable refers to the middle of an
instruction encoding instead of the start of an instruction.

R=christofferqa@google.com

Bug: 110164501
Change-Id: I4c37ec45112f0f8f9a203cdc4ac744049a95cbe3