Avoid <= in version checks as it is then easy to miss MR releases.
We had a check for Lollipop VMs that was checking that the
min sdk version was <= L(21). However, that is incorrect, as
L_MR1(22) is also Lollipop and has a larger sdk version.

Updating to always use < on the next version number so that
we check that the min sdk version is < M(23) which then
includes all MR releases.

R=christofferqa@google.com

Bug: 111146567
Change-Id: I4ec2b5ec22bb63a0d5753603958edc34ae841e28