Fix local variable table in CF backend when inserting CfFrames.
In certain cases we would end up extending the variable scope
across a CfFrame instruction in the CfCode even though the
CfFrame is not in the scope of the variable. This leads
to CfCode that we cannot turn into IR because of undefined
values during IR processing.

In particular, this situation occurs when there is a return
followed by a move-exception instruction in the IR. We have
to generate a frame for the move-exception entry, but we
need to make sure to end the locals at a label before the
CfFrame.

R=christofferqa@google.com, jsjeon@google.com, tamaskenez@google.com, zerny@google.com

Bug: 76181966, 111619623
Change-Id: Icc372f270a271579b55c08866ff1db56b1969881