Update AdaptResourceFileContentsTest to use ArchiveResourceProvider
This CL removes the implementation of getDataResourceProvider() from ArchiveProgramResourceProvider, since it is expected to return null in the AGP integration.

This method was indirectly used by AdaptResourceFileContentsTest.testProguardBehavior() to retrieve the set of resources in a jar; this CL therefore also updates the test so that it retrieves the resources of a jar via other means.

Change-Id: I97e2e397235a61d67b45f5ea91734372c6988be9