Insert local changes between blocks in split edges.
This CL refactors the IRBuilder and SourceCode companions to have a buildBlockTransfer callback which is responsible for emitting the local variable changes between blocks.

In addition, the use of debug-value reads is now only needed explicitly where a local variable ends, which can be:
1. between two instructions within a basic block (either it just ends, or the local is written to)
2. between two basic blocks, in which case it now always ends prior to entry of the successor, this happens in one of three ways:
 a. if only one successor exists, by updating the locals prior to exit of the block; or
 b. if multiple successors exist, for each successor:
   - if the successor has only this as predecessor, emit the local change on entry, but with the position of the predecessor;
   - otherwise, insert a split block and closing the local in the split block;

A local end ensures that the local remains live until the end and is then used to explicitly end the liveness at that point (even if the SSA value may continue being live after that point).
For this reason, whenever a local should start, we now must emit a DebugLocalWrite introducing it even if the source of the write is an SSA value associated with the same local, since it may have been previously ended.

Bug: 80385846
Bug: 78611610

Change-Id: I94e4ac1f00a0b849f29d727a2509376858519808