Introduce relaxed method signature equivalence.
This equivalence is in between MethodJavaSignatureEquivalence
and MethodSignatureEquivalence: it is stronger than the former in that
it considers the return type of methods; it is weaker than the latter
in that it allows covariant return type: a method in a sub type can
return a narrower type.

Method minifier uses this equivalence for non aggressive renaming,
which still satisfies multi interfaces and lambda renaming that
require fine-grained method wrapping as well as satisfies the case
with covariant return type that requires less strict wrapping.

Bug: 112185748, 76461545
Change-Id: Ib462dbd81e598e4b11b7f787da4f5cfe3e3bca3b