Mark argument types as live.
We assume that argument types would be seen and marked as live while
walking through live methods, as long as arguments are used. If some
argument is not used, an instance of any type can be flown, resulting in
missing argument types. Unlike JVM, which checks types when instances
are actually used, ART checks type compatibility of argument/parameter
when invocation target is identified, hence class not found error. It is
safer to mark the type of arguments as live when we handle live methods.

Bug: 112452064
Change-Id: I83c4c01740127b38214f7a59ce094fd316fa4d4c