Support for $closeResource synthesized by java 9 compiler
Pilot implementation of rewriting $closeResource(...) methods
synthesized by java 9 compiler for some cases of try-with-resources.

The problem with current implementation is that on pre-19 devices
even though java.lang.AutoCloseable is available in platform (not in
android.jar), is not fully supported in a sense that not all classes
implement it properly.

We have to replace it with our version of $closeResource(...) which
takes the resource as java.lang.Object and checks if it can cast
it to AutoCloseable.

To avoid duplicating the methods we also use one single method defined
in an utility class '$r8$twr$utility' created when needed. This class
needs to be deduped for incremental dexing scenarios.

Bug: 70292540
Change-Id: I4da89a8482ab6895267e58c3debc7ad4e12a3dfe