Heuristic for inlining into blocks with catch handlers
This CL avoids inlining a method when the call site has one or more catch handlers, and the method being inlined has several instructions that may throw. This prevents that inlining leads to an explosion in move-exception instructions and resolutions moves that are needed to setup the state for the catch handlers.

This saves approx. 30kB on gmscore-v9 and 169kB on gmscore-latest. The combined savings of this CL and https://r8-review.googlesource.com/c/r8/+/24847 are approx. 76kB for gmscore-v9 and 120kB for gmscore-latest.

Bug: 64432527
Change-Id: Iee115a9765e31bb8322817ca075154834d9773e5