Detect the naming loop due to -useuniqueclassmembernames.
For methods with same signature but different return type, method
minification can be stuck while trying the same rename candidate over
and over again if -useuniqueclassmembernames is set.

Here is the dilemma:
That option forces the naming state to return the same renamed name for
the same signature---method name and parameters. If we're using Java
signature to wrap those methods, they will be grouped together, hence
renamed together w/o any problems when that option is on.

However, when it comes to assigning such renamed names to methods in
implementing subtypes, using Java signature made it next to impossible
to resolve which method in implementers corresponds to which method in
interfaces (if multiple unrelated interfaces have methods with same
signatures). Thus, we have to use a method equivalence stricter than
Java signature. (See http://b/76461545 for more details.)

------

One stopgap is to detect such case and break the naming loop.
We don't need to collect what rename candidates have been tried. If
-useuniqueclassmembernames is involved, the same candidate should be
tried again and again, so checking two consecutive trials is enough.

That candidate would be regarded as `not available` as the other,
unrelated interface method is already using that name, while the option
forces us to assign that renamed name, hence loop. :$
The reason they're regarded as different naming states is to assign
consistent names in the same hierarchy. Because they're unrelated,
using that seemingly conflict candidate is safe yet satisfying everyone:
we still don't violate -useuniqueclassmembernames;
we escape from the naming loop; we can resolve which methods implement
which interface methods; and thus we can consistently assign the same
renamed name to all the involved methods.

Bug: 112701848
Change-Id: I13b59b4b540ab61928d602622cd68531da0b4a5c