Remove special case for synthetic constructor argument types.
The argument and return types are now always marked as live for
all method invocations that end up in the result to ensure that
the result verifies.

Parameter and return types are marked live for all targeted as
well as all live methods. There can be targeted methods that are
not live as well as live methods that are not targeted and
therefore we need to handle parameter and return types in both
cases.

R=sgjesse@google.com

Change-Id: Ic8f7d89fc330721a57701d5b6e44e0ec3b15d7dd