Workaround massive dex2oat memory usage for self-recursion.
On Marshmallow, dex2oat can take up massive amounts of memory
when processing self-recursive methods due to inlining. Having
a try-catch disables inlining. Therefore, when targeting
Marshmallow or earlier, we add a catch handler to methods with
more than 4 self-recursive calls to disable inlining in dex2oat.

There is no guarantee that this will remove the out of memory
situations on all devices. However, it should help limit the
issue.

R=sgjesse@google.com, zerny@google.com

Bug: 111960171
Change-Id: I226a621e1dc0cea0f0b844f1cc9ebb5c3f579763