Change collection of call sites for minification
For minification the call sites used in the program must be known
to ensure correct renaming of multi-interface lambdas.

This changes how we collect call sites:

* Remove the canonicalization of call sites in DexItemFactory
* Collect live call sites during tree shaking
* Collect all call sites that has been desugared

Use the union of all collected call sites during minification.

The warning for missing interfaces is turned off during
minification. We don't prune the collected desugared call sites,
so there might be call sites for which the interface does not
exist in the final output.

Bug: 113100603
Change-Id: Ib3a5aefa0f29bc8d0684334dada7ef3f4b42b666