Enable CF on ClassInlinerTest, change in rewriteMoveResult:
Previously, for CF backend, rewriteMoveResult was a no-op because with JVM
byte code we can't suppress the result value of a function call.

The test revealed that not recognizing that an input argument (e.g. the
receiver) is the same as the output value introduces unnecessary users for that
invoke which prevents class-inlining later.

This CL changes rewriteMoveResult for CF such that we do the same as with DEX
except we retain the outValue.

Change-Id: I5ac9d11902c355fc87fe6d5cb94be18a3ba34f09