Attach type lattice to value and deprecate TypeEnvironment.
Due to cyclic phis, the initial analysis for a given method should use
fixed-point iteration: repeatedly updating value/phi's type lattice.
By definition, phi's type is a least upper bound of all operands, hence
its type lattice is moving upward: something wider. That phi type could
be flown to other values, so in general, during the initial analysis,
everything could be widening.

After that, by inlining, we might be able to see more specific info,
which could narrow the type lattices. One caveat is that inlinee's IR
is built on the fly, so when to widen/narrow the type environment for
the inlinee is tricky. Newly generated inlinee blocks need to be
widen, while the return value of the inlinee, which might have a
narrower info, should be flown to the original context separately.

------

In this CL, some optimizations, e.g., inliner, devirtualizer, non-null
tracker, are updating type lattice accordingly, but some are still not
doing so. Thus, asserting that type lattice changes is monotonic cannot
be enabled yet. The best places to initialize/update type lattices are:
when IR is built, when Value is newly created; when operands in phi are
added, replaced, or updated. After that, which will be a follow-up CL,
optimizations don't need to update type lattices; plus, can use
type/null info that is attached to value by construction.

Bug: 72693244
Change-Id: Ie0e8218ae4316bab671852d3b58654999da9a673