Do not alter already native/abstract methods while tree pruning.
If native method is targeted but not live, TreePruner regards that as
non-abstract, and adds throwing instructions. That results in dex2oat
error: a method has code, but is marked native or abstract.

Also, a couple of refactorings:
+ introduced hasCode() and shouldNotHaveCode() into DexEncodedMethod
and applied them to relevant use points.
+ assert we're not making final method _abstract_, which is found
while trying a little bit different approach.

Bug: 113138046
Change-Id: I04da7d4e56aed836376824efa3cba530b1522cd3