Make our code forward compatible with fastutils 8.2.1.
In fastutils 8.2.1 there are forEachRemaining versions that take
a consumer accepting and int or a consumer accepting an Integer.
We therefore have to disambiguate. For now we use the Integer version.
When we can update to 8.2.1 we will get deprecation warnings and can
move to the int version that doesn't box all these integers.

R=sgjesse@google.com

Change-Id: I3f2f336786a9d8b470c85d825dae6922b39ad01c