Workaround mediatek constructor inlining bug.
Some mediatek lollipop VMs have a peculiar bug where the inliner
crashes if there is a simple constructor that just forwards its
arguments to the super constructor. Strangely, this happens only
for specific signatures: so far the only reproduction we have is
for a constructor accepting two doubles and one object.

To workaround this we insert a materializing const instruction
before the super init call. Having a temporary register seems
to disable the buggy optimizations.

R=sgjesse@google.com, zerny@google.com

Bug: 68378480
Change-Id: I2e587393f10cc2b1200fb551d9646c004afe43e5