Add easier control of whether data resources are processed or not by R8
This is only relevant for R8. D8 will always ignore data resources.

When the current BaseCompilerCommand.setOutput API is used to specify
the output, there is no way of controlling if data resources are
processed or not. Before this change the consumers created when this
API was used would ignore data resources for R8 as well. The only way
to process data resources with R8 was to use a custom consumer through
the API.

This adds an additional setOutput method to the R8Command API which
takes a boolean argument to control if data resources are processed or
not.

The default when the original setOutput is used is now to process data
resources for R8.

Added the option --no-data-resources to the command line to be able to
override this default behaviour.

Change-Id: I87d30c99133bfc365ec30eaacfa1c0755acf38e1
Bug: