Apply proto transformer before name-lookup in minifier
This CL fixes a bug in the minifier that caused the implementation of an interface method to be renamed to another name than the interface method it was implementing (leading to an AbstractMethodError at runtime). This issue only arose when covariant return types were being used.

The problem is fixed by applying the proto transformer (that drops the return type of the proto if aggressive overloading is not enabled) before looking up to see if a method has already been renamed in a supertype, and by relaxing MethodSignatureRelaxedEquivalence to ignore the return type entirely.

Due to the changes to MethodSignatureRelaxedEquivalence, we will now rename all methods that have the same name and and the same parameters to the same name (see CovariantReturnTypeTest).

Bug: 114554345, 115613850
Change-Id: Id7a42cc4a23e7ebf54adc071e1a65f3315baf537