Adjust keep rules for atomic field updaters
This CL adjusts the enqueuer such that fields that are passed to Atomic*FieldUpdater.newUpdater() will be kept independent of whether or not their holder or one of its subtypes are instantiated. This is necessary to prevent a NoSuchFieldException from the call to newUpdater(), since the method checks that the field is present upon creation of the updater.

Bug: 114554345
Change-Id: I17ff40dd29c092065c4ae5c5021e3f0bbacae9e8