Update synthetic bridges after vertical class merging
This CL fixes an issue in the vertical class merger that has to do with synthetic bridge methods that get inserted during class merging.

The problem is that these bridge methods are created before vertical class merging is finished (meaning that the signatures of the bridges may change from the time they are created until vertical class merging finishes), and that the graph lense does not have any mappings for these intermediate method signatures". This CL therefore processes all the synthetic bridge methods when all classes have been merged.

Example: Consider that a virtual method "void A.foo(B)" is merged into the subtype of A. When that happens, a bridge method "void ASub.foo(B)" is created, which invokes the method "void ASub.foo$A(B)" via an invoke-direct instruction. After this bridge method has been created, the type B may be merged into its subtype, but the signature of the bridge method continues to be "void ASub.foo(B)". This is problematic, since the graph lense only provides capabilities for mapping back and forth from the end-signature "void ASub.foo(BSub)".
Change-Id: I6a65ecd103fe017d295a77c44f29cce4a17789a1