Keep a class if dependent rules keep instance members
When a class is live (mentioned in a program) and a dependent rule
(e.g. -keepclassmember) specify instance members, R8 will now also
keep the class.

Before only the dependent static members where kept - unless the class
was seen as instantiated - in which case the dependent instance
members was also kept.

The reasoning is: If a dependent rule specify instance members this
indicate that this class will have instances - no matter if there are
instantiations directly in the code.

This costs 51KB on GmsCore v9.

Bug: 115867670
Change-Id: I2c56327a70e29b7f2d3c6b5bf79c02c33f66841e