Improve collection of constant arrays for reflective calls
In reflective calls like this

  Method getMethod(String name, Class<?>... parameterTypes)

R8 will try to find invocations where name and parametertypes
arguments are a statically known, in order to apply renaming to these
invocations.

The previous implementation only supported collection array content
for parameterTypes when the construction of the array and its content
was in the same basic block. This was not able to determine the array
content if the reclective call was inside a try block.

With this changes R8 will follow all array operations from the
array construction to the use at the reflective call site. This is
conservative and will bail out if the array content cannot be
statically known.

Bug: 116078230
Change-Id: I7712c2e7fecafda645f2802b694ac80ca50844f3