Keep a class if conditionally dependent rules keep instance members.
This is an -if version of workaround for b/115867670: if dependent rule,
e.g., -keepclassmembers, specifies instance members, we need to keep the
holder when it is live, assuming the holder could be instantiated. The
same situation holds for -if conditional rule.

To piggyback on the existing workaround, members mentioned at
-keep/classmembers are passed to Enqueuer via dependentNoShrinking.
Then, the same workaround, which visits dependent items and enqueue
the holder if rules for non-static members are found, is applied too.

------

This CL also fixes access modifier and annotation handling in -if rule.

Bug: 116092333
Change-Id: Ib5e9540b383394db15cfc9a24b0c40dad2149bb5