Handle non-standard <init> calls in class inliner and merger
This CL aborts class inlining in case of the following code pattern (where B is a subtype of A), instead of failing with an assertion error:
  new-instance v0, B
  invoke-direct A.<init>, v0

Furthermore, in case A is merged into B by the vertical class merger, compilation fails with an error telling the user to add a -keep rule for class A. The motivation for this is that the vertical class merger cannot easily recognize the above code pattern, since it runs prior to IR construction. Therefore, we currently allow merging A and B although this will lead to invalid code, because this code pattern does generally not occur in practice (it leads to a verification error on the JVM, but not on Art).

Bug: 116282409
Change-Id: Ia57badc8cb52e3164421ddd908918c7b4699e415