Extend the liveness of the receiver to the entire method.
This is necessary to avoid Art crashes when there is a
debugger attached to break on exception in a release build.
Art will crash if asked for the |this| object in a JIT
compiled method where the receiver register has been
clobbered.

We were already doing this live range extension for anything
with min-sdk that includes Lollipop devices because the
verifier had a similar assumption about the receiver
register not being clobbered back then. Therefore,
this should not be visible to most users.

This loses 5kB on GMSCore when comparing R8 compilations
with and without the extension of the live range.

R=sgjesse@google.com, zerny@google.com

Bug: 116837585
Change-Id: Id06f745fd5a540d26b89e13f8d503dcb95a0a4a9