Ensure that we only sort members synchronized Additionally, avoid copying and sorting the same methods twice
There was a race condition where we could sort the directmethods
of a synthesized class that we needed in several dex output files in one thread,
while cloning them in another.

We now sort the methods before adding them to the collector,
avoiding a copy and sort (which we will do on the same methods
right after anyway)
The sorting is done thread safe.

Change-Id: If2eb0ac2af61af20f26e86840a919325206e4d81