Add regression test for crash during debugging in certain Art VMs.
As expected it is a little complicated to consistently reproduce
this issue. In order to make sure that the JIT kicks in we have
to make sure not to set breakpoints but instead just install
"break on exception". Or debugger tests did not have support for
that so this change adds it. The added test crashes Art 7.0.0
in the expected way with out the workaround.

R=zerny@google.com
Bug: 116837585

Change-Id: Ifca8e267b2b6d2cb7beb8bb71b7858f52e82a44f