Relax assertions during check-cast removal
This CL relaxes the following two assertions from the check-cast removal:
  assert outTypeLattice.equals(inTypeLattice);
  assert outTypeLattice.equals(castTypeLattice);

These asserts are overly restrictive as they require that the two types agree on the nullability, although they don't need to. For example, it is allowed that the in-type is not nullable and the out-type is nullable. This is a loss of precision, but still sound.

For convenience and clarity, this CL introduces a simple null-lattice that can be used for the soundness checks.

Change-Id: I098eb77be2eb032a6b8533851080ee90995040eb