Extend reflective analysis to include java.lang.Enum.valueOf
Whenever valueOf on an enum is used (either directly on an enum type
of through java.lang.Enum.valueOf) the library code invokes the method
values() on the enum class.

This change ensures that the values() method stays and is not renamed
when it can be determined that valueOf is used on a particular enum.

Bug: 117299356
Change-Id: Id12437ccfc84aa6c04924fc59d27b99ee6dee689