Remove class hierarchy checks that all abstract methods are implemented.
If we have a number of large sub class hierarchies of an abstract
class but an abstract method is only ever called with a receiver
of some of these sub class hierarchies, it is valid to remove
the implementation of the abstract methods from the classes
in the other sub class hierarchies (as they can never get
called at runtime).

See the Invoke.getType() class hierarchy in R8 for an example.
It is valid to remove the InvokeCustom, InvokeNewArray and
InvokeMultiNewArray versions of getType as they can never be
called.

R=christofferqa@google.com, sgjesse@google.com

Bug: 115705526
Change-Id: I35dbdf5ee95f16649713aa1e19c4c6d95fd2faba