Implement workaround for Dalvik tracing JIT tracing past the end of instructions.
The Dalvik tracing JIT can trace right past a throw instruction and into
the next bytes whether or not that is code. Tracing off the end of
instructions lead to crashes. When targeting Dalvik VMs we insert
"nop; goto -1" after a throw if it is the last instruction in a
method with backwards branches.

R=zerny@google.com

Bug: 117907456
Change-Id: Iacdcd5a8fe666ebc87a58dc1ed92ba701f274260