Compute highestSortingString in DexCode.collectIndexedItems()
This moves the computation of DexCode.highestSortingString to DexCode.collectIndexedItems(). This is needed because R8 is allowed to change field, method, and type names, and it is therefore not safe to compute the highestSortingString during parsing.

This change also ensures that DexCode.highestSortingString cannot be a DexItemBasedString. This is important because DexItemBasedString breaks canonicalization. A subsequent CL should introduce a DexItemBasedConstString instruction and remove DexItemBasedString, to ensure that DexString is canonicalized entirely.

Bug: 117828645
Change-Id: I82753e117895f493e57e28b5e7e44d489b7aabc8