Compute non-null parameter hints in general.
This CL piggybacks on checkNullReceiverBeforeSideEffect, which checks
if the current code is preserving NPE semantics after inlining certain
callee's bodies. That walks through blocks; marks the block if it
encounters an instruction that may throw NPE for the receiver;
makes sure that NPE is triggered before any other side effects.

When it comes to non-null paramter hints, we want to mark a parameter is
non-null when the code reaches normal exits. That is, if a parameter is
null, the code will raise NPE somewhere, and thus we can use the above
routine with little modifications: it previously dealt with instruction
types only, but we want to handle a specific code pattern:
  if (param == null) throw NPE

This will allow us to compute non-null param hints for
Instrinsics#checkParameterIsNotNull, unblocking inlining threshold
(b/117848700). We can do more sophisticated analysis (b/71500340).

------

For kotlinc-generated class files, we will still use two shortcuts:
1) metadata info where parameter type can have a clear indicator, like
String? v.s. String. 2) If metadata is stripped off somehow, we can
still rely on the fact that, as a safety guard of inter-operability,
Intrinscis#checkParameterIsNotNull is used to check the nullability of
parameters.

Note that, Intrinscis#checkParameterIsNotNull doesn't have metadata
we can extract, and cannot use the 2nd shortcut as well since it does
not call itself (to check param nullability).

Bug: 117848700, 71500340
Change-Id: I512db16ec9d6f3d5bc5bb4be360fa3d8a85d9df1