Add initial support for removing unused arguments
This does a conservative collection of unused arguments for static
methods. The conservative approach look through the Java byte code for
arguments which are definitely not used.

The lens code rewriting have added support for removed arguments as
well.

The optimization is turned off by default and only enabled for the
tests for now.

This saves about 5 Kb on GMS Core v10.

Bug: 65810338
Change-Id: I46bff039b0568f52ae15f1be72b0086dea6dbb20