Use same matching semantics for fields and methods rules
For methods mentioned in a keep rule all matching methods on all
classes in the super chain are included when building the root set.

However for fields this was not the case. Only fields specifically on
the classes matched by the rule was included.

This is now changed so that matching for fields and classes follow the
same rules. By reflection it is always possible to get to fields (as
well as methods) in the super chain, independent of access
modification and shadowing.

Bug: 118851616
Change-Id: Ief099afc84bf1f8c2d4bcb8996722f24b74e9c5c