Verify correctness of types
This CL extends IRCode.verifyTypes() such that we now check that all in-values to field-put instructions are a subtype of the corresponding field type. Assignments where the type of the in-value is based on a missing type are ignored.

One issue is that the input program may not type check. This should generally not happen, but it could happen if the input has been optimized and the optimizer has a bug.

To avoid assertion failures from calls to IRCode.verifyTypes() on code that was already broken in the input, this CL introduces a type checker that runs immediately after lens code rewriting. If a method cannot be type checked, the code is assumed to be unreachable and a warning is emitted.

Bug: 119042956, 119181813
Change-Id: I51d1e21edc606e04f4c79679b110ca5a4360fea4