Revert part of "Change the semantics for keeping methods"
This reverts part of commit de3dab0050043c50c5c6879fc8517fda071c31ce.

The change to Enqueuer.markMethodAsKept was wrong, and caused
desugaring of default methods to break. The change was not needed
after all. The changes to the RootSetBuilder was sufficient.

Added a more extensive test of default methods.

Bug: 118851616
Change-Id: I03c7ea8ee646d318f17df5a9bad97b8e34b1a18a