Fix mapping to original field/method signatures after class merging
When vertical class merging has been run, and it is known which classes that have been merged, a fix-up phase is responsible for changing all the field/method signatures that refer to a class that is no longer present due to class merging.

However, the GraphLense builder does not record the mapping from the new signatures to the old signatures, which means that the getOriginalMethodSignature() and getOriginalFieldSignature() do not work as intended after class merging.

Change-Id: I8f107b2824346c4b3941ad6988f35e39188f9434