Add support for class name reflection with minification.
For to-be-renamed class's name computation, we replace name reflection
with DexItemBasedConstString, along with what kind of name computation
is needed. Also, propagate useIdentifierNameString marker to indicate
the method has artificial DexItemBasedConstString for name reflection.

One caveat is, for getSimpleName, we should know whether the target
class is local or member class. To do so, we need to retrieve inner
name, but if .class.getSimpleName() is the only usage of that class,
replacing it with DexItemBasedConstString make that type not live.
This CL adds necessary changes to registry logic to trace type reference
inside DexItemBasedConstString.

Bug: 118536394
Change-Id: I3c76c1c33fda2d8e0fda59b911bd3f01acd0af51