Add main dex list consumer api to D8.
When merging classes for a multidex build, D8 honors the input
main dex list and also puts anything that is synthesized because
of main dex classes in the main dex file.

When building app bundles, the main dex list is put in the bundle
so that feature split code can be merged with base split code
for shipping to pre-L devices. However, since the main dex list
contains no synthesized classes, bundletool can generate an apk
with some needed classes missing from the main dex file.

In order to fix this, D8 should produce an updated main dex list
containing all the classes that it synthesized and determined
were needed in the main dex file.

R=gavra@google.com, zerny@google.com

Bug: 120039166
Change-Id: I6dd94f8c3ed8617b89fbd1c18ec69bca27172437