Add script for archiving gradle test logs
Move archive functionality to utils.py.

Create explicit script for archiving.

I will change the bots to run this whenever we fail on the bots, to also handle the
timeout case where the bots kill the test.py script (in which case we currently do not
get the logs archived)

I will do a follow up change to remove the archiving logic from test.py when this is
on all the bots

Change-Id: Iabcba8992d60deac1cfa87b1cca0c47d6d493d15