Don't warn for main dex classes which are removed during tree shaking
When the main dex list is computed using main dex rules don't give
warnings about classes expected to be in the main dex which was later
removed during tree shaking.

We still give warnings for classes in an explicitly provided main dex
list which are not present in the output.

Bug: 119935815
Change-Id: I3ef1e74ef13dcbf288c8d604e4bc8953d6f2e790