Reland "Order outputs when writing to a zip"
This will ensure that we write the classes*dex files in the same order
and that we write any resources passed through also in the same order.

This cl includes:
Interleave directories with files in archives

This should provide a more readable ordering in the zip file.

Original reviews:
https://r8-review.googlesource.com/c/r8/+/31381
https://r8-review.googlesource.com/c/r8/+/31400

Reverted here (but not the problem):
https://r8-review.googlesource.com/c/r8/+/31511

Bug: 119945929
Change-Id: I94f96373a850e233ae9a75b6563dd38ada7d7619