Add a script for building Android Studio apps
The Android Studio app will be checked out, build and signed with R8,
R8 full mode and Proguard.

The Android Studio project has to be modified to build an unsigned apk
in release mode, and a checked-in keystore file will be used for the
signing with apksigner from the Android SDK build tools.

The three signed apks will be in build/<appname>out/.

Configuration included for one app 'tachiyomi'.

Location of the Android SDK is hardcoded to $HOME/Android/Sdk.

Bug: 119935815
Change-Id: Icf912bb2f6ea3c318622eddc5a9a3dbc87d449c4