Return unmodifiable list in BasicBlock.getPredecessors() and BasicBlock.getSuccessors()
The purpose of this change is to be able to determine if any changes has been made to the graph. With this change, this can now be done conservatively by checking if anyone has asked to get the mutable successor/predecessor list. This will enable us to assert that no obsolete dominator trees are being used.

Bug: 120114190
Change-Id: I4bf3ff321dce6a4aa573fca4f452180f1c6794f5