Support passing in a main dex list to the dex splitter
This will be used when generating the dex files for the base split.

If none of the splits are spliced into the pre-L generated apps by
bundletool (which already uses the main dex list) the dex is simply
copied over. This means that we need to guarantee that whatever is
in the main dex list is also used for generating the base split.

Bug: 120845002
Change-Id: Id48e98d667459d41addf74e0db0103580197a9e0